requirements.txt
openai
google-generativeai
streamlit
pandas
PyPDF2
